+++ src/lib/reseau_graphe/ospf/PaquetLSR.hpp
@@ -15,22 +15,22 @@
 class PaquetLSR: public PaquetOSPF {
     private:
         // Attributs
-        uint8_t m_IdEmetteur;
+        uint16_t m_IdEmetteur;
         std::vector<std::bitset<32>> m_IdLSADemandes;
 
     public:
         // Constructeur
-        PaquetLSR(const uint8_t& idEmetteur,
+        PaquetLSR(const uint16_t& idEmetteur,
                   const std::vector<std::bitset<32>>& idLSADemandes);
 
         // Destructeur
         ~PaquetLSR();
 
         // Getters et setter
-        void setEntete(const TypePaquet& type, const uint8_t& idRouteur) override;
+        void setEntete(const TypePaquet& type, const uint16_t& idRouteur) override;
         const TypePaquet& getType() const override;
-        const uint8_t& getIdRouteur() const override;
+        const uint16_t& getIdRouteur() const override;
 
-        const uint8_t& getIdEmetteur() const;
+        const uint16_t& getIdEmetteur() const;
         std::vector<std::bitset<32>>& getIdLSADemandes();
 };
